backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.0.1  
passlib[bcrypt]==1.7.4
boto3==1.34.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.1.0.post1